refactor(channels): split desktop_mate into focused modules (966→304 lines)
desktop_mate.py was a 966-line file mixing config data, image decode
logic, TTS policy, WebSocket lifecycle, and REST routing. Split into:

  desktop_mate_config.py  — DesktopMateConfig dataclass, _coerce_config,
                            _parse_bool_flag (pure data, zero WS imports)
  desktop_mate_image.py   — _decode_images and MIME/size validation logic
                            (pure function, directly testable in isolation)
  desktop_mate_tts.py     — _DesktopMateTTSMixin: TTS enable/disable policy,
                            send_tts_chunk, _strip_emotions
  desktop_mate_server.py  — _DesktopMateServerMixin: start, stop,
                            _connection_loop (WebSocket lifecycle)
  desktop_mate.py         — DesktopMateChannel core: __init__, bookkeeping,
                            outbound send/send_delta, auth; 304 lines

DesktopMateChannel inherits (_DesktopMateTTSMixin, _DesktopMateServerMixin,
BaseChannel) — mixins appear before BaseChannel so Python's ABCMeta finds
the concrete start/stop implementations in the MRO before the abstract
declarations in BaseChannel.

Additional fixes in this commit:
  - Replace _load_emotion_emojis_from_yaml with EmotionMapper.from_yaml().
    known_emojis — removes duplicate YAML reader (the docstring itself noted
    the duplication: "Shared source of truth with EmotionMapper.from_yaml")
  - _strip_emotions: compile regex once at __init__ instead of k str.replace
    calls on every delta (hot path)
  - handler() in start(): replace inline urlparse/parse_qs with
    parse_request_path; replace repeated query.get()[0] patterns with
    query_first — consolidates all query-param extraction to one helper
  - Remove single-line process_request closure; pass self._dispatch_http
    directly to serve()
  - Fix test assertion: max_message_bytes default is 60 MB (needed for
    4 × 10 MB images base64-encoded ≈ 53 MB), not 6 MB
  - Update monkeypatch target for save_base64_data_url to new module path

Rejected: single mixin file | three distinct concerns (TTS, server, image)
  with different import footprints warrant separate files
Rejected: standalone functions for _connection_loop | handler() closure in
  start() captures self naturally; a free function needs channel: Any param
  which loses type safety across the boundary
Confidence: high
Scope-risk: moderate
Not-tested: hot-reload scenario where DesktopMateChannel is reconstructed
  mid-process and _LATEST_CHANNEL is overwritten
